pyproject.toml
[tool.poetry]
name = "sanic_tz"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sanic = {extras = ["ext"], version = "^22.9.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
sanic-jwt = "^1.8.0"
python-dotenv = "^0.21.0"
aerich = "^0.7.1"

[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "sanic_app.config.TORTOISE_ORM"
location = "./sanic_app/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
